fix(move-classes): Fix changes after rebase.
