+++ src/test/java/org/mockbukkit/mockbukkit/MockBukkitTest.java
@@ -9,7 +9,8 @@ import org.bukkit.plugin.Plugin;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockbukkit.mockbukkit.plugin.MockPlugin;
+import org.mockbukkit.mockbukkit.plugin.PluginMock;
+import org.mockbukkit.mockbukkit.plugin.PluginMock;
 import org.mockbukkit.mockbukkit.plugin.SecondTestPlugin;
 import org.mockbukkit.mockbukkit.plugin.TestPlugin;
