+++ src/test/java/org/mockbukkit/mockbukkit/scheduler/paper/FoliaAsyncSchedulerTest.java
@@ -3,7 +3,7 @@ package org.mockbukkit.mockbukkit.scheduler.paper;
 import org.mockbukkit.mockbukkit.MockBukkit;
 import org.mockbukkit.mockbukkit.MockBukkitExtension;
 import org.mockbukkit.mockbukkit.MockBukkitInject;
-import org.mockbukkit.mockbukkit.PluginMock;
+import org.mockbukkit.mockbukkit.plugin.PluginMock;
 import org.mockbukkit.mockbukkit.ServerMock;
 import org.mockbukkit.mockbukkit.scheduler.BukkitSchedulerMock;
 import org.junit.jupiter.api.BeforeEach;
